requirements.txt
openai
pandas
python-dotenv 
google-generativeai 
google-ads